Create /interfaces folder in odoo module
