+++ src/modules/odoo/application/interfaces/field.interface.ts
@@ -1,0 +1,7 @@
+export interface IField {
+  id: number;
+  name: string;
+  model: string;
+  selection: false | string;
+  selection_ids: number[];
+}
