+++ src/modules/odoo/application/interfaces/model.interface.ts
@@ -1,0 +1,5 @@
+export interface IModel {
+  id: number;
+  name: string;
+  model: string;
+}
