+++ src/modules/odoo/application/interfaces/order-line.interface.ts
@@ -1,4 +1,4 @@
-export interface IOrderLineResponse {
+export interface IOrderLine {
   product_id: [number, string];
   product_uom_qty: number;
 }
