+++ src/modules/odoo/application/interfaces/sale-order.interface.ts
@@ -1,0 +1,3 @@
+export interface ISaleOrder {
+  order_line: number[];
+}
