--- src/modules/odoo/application/responses/sale-order.response.interface.ts
@@ -1,3 +1,0 @@
-export interface ISaleOrderResponse {
-  order_line: number[];
-}
