--- src/modules/odoo/application/services/odoo.models.ts
@@ -1,5 +1,0 @@
-export enum MODEL {
-  PRODUCT = 'product.template',
-  SALE_ORDER = 'sale.order',
-  ORDER_LINE = 'sale.order.line',
-}
